tccWebapp/requirements.txt
django
django-crispy-forms
crispy-bootstrap5
dotenv
psycopg2-binary
pillow